Split advisory upstream client protocols
