+++ src/app/services/advisory_policy_service.py
@@ -1,7 +1,7 @@
 from typing import Any
 
 from app.contracts.advisory_policy import AdvisoryPolicyEnvelopeResponse
-from app.services.upstream_client_protocols import AdvisoryPolicyClient
+from app.services.advisory_client_protocols import AdvisoryPolicyClient
 from app.services.upstream_envelope import build_gateway_envelope, raise_for_upstream_error
