+++ src/app/services/upstream_client_protocols.py
@@ -77,573 +77,6 @@ class LotusAiWorkflowClient(Protocol):
         correlation_id: str,
     ) -> tuple[int, dict[str, Any]]: ...
 
-
-class BankDemoProofClient(Protocol):
-    async def get_bank_demo_proof_scenario_contract(
-        self,
-        *,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_bank_demo_supported_claim_register(
-        self,
-        *,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def build_bank_demo_proof_pack(
-        self,
-        *,
-        body: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-
-class AdvisoryWorkspaceClient(Protocol):
-    async def create_advisory_workspace(
-        self,
-        *,
-        body: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_advisory_workspace(
-        self,
-        *,
-        workspace_id: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def apply_advisory_workspace_draft_action(
-        self,
-        *,
-        workspace_id: str,
-        body: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def evaluate_advisory_workspace(
-        self,
-        *,
-        workspace_id: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def save_advisory_workspace(
-        self,
-        *,
-        workspace_id: str,
-        body: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def list_advisory_workspace_saved_versions(
-        self,
-        *,
-        workspace_id: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_advisory_workspace_saved_version_replay_evidence(
-        self,
-        *,
-        workspace_id: str,
-        workspace_version_id: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def resume_advisory_workspace(
-        self,
-        *,
-        workspace_id: str,
-        body: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def compare_advisory_workspace(
-        self,
-        *,
-        workspace_id: str,
-        body: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def request_advisory_workspace_rationale(
-        self,
-        *,
-        workspace_id: str,
-        body: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def review_advisory_workspace_rationale(
-        self,
-        *,
-        workspace_id: str,
-        body: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def handoff_advisory_workspace(
-        self,
-        *,
-        workspace_id: str,
-        body: dict[str, Any],
-        idempotency_key: str | None,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-
-class AdvisorCockpitClient(Protocol):
-    async def list_advisor_cockpit_actions(
-        self,
-        *,
-        params: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def list_advisor_cockpit_preparation_packets(
-        self,
-        *,
-        params: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_advisor_cockpit_action(
-        self,
-        *,
-        action_item_id: str,
-        params: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_advisor_cockpit_snapshot(
-        self,
-        *,
-        params: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_advisor_cockpit_supportability(
-        self,
-        *,
-        params: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def acknowledge_advisor_cockpit_action(
-        self,
-        *,
-        action_item_id: str,
-        body: dict[str, Any],
-        params: dict[str, Any],
-        idempotency_key: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def evaluate_advisor_cockpit_house_view_cohort(
-        self,
-        *,
-        body: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-
-class AdvisoryPolicyClient(Protocol):
-    async def list_policy_packs(
-        self,
-        *,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_policy_pack_version(
-        self,
-        *,
-        policy_pack_id: str,
-        policy_version: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def validate_policy_pack_version(
-        self,
-        *,
-        policy_pack_id: str,
-        policy_version: str,
-        body: dict[str, Any],
-        idempotency_key: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def activate_policy_pack_version(
-        self,
-        *,
-        policy_pack_id: str,
-        policy_version: str,
-        body: dict[str, Any],
-        idempotency_key: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def create_policy_evaluation(
-        self,
-        *,
-        proposal_id: str,
-        proposal_version_id: str,
-        body: dict[str, Any],
-        idempotency_key: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_policy_review_queue(
-        self,
-        *,
-        evaluation_status: str | None,
-        portfolio_id: str | None,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_policy_evaluation(
-        self,
-        *,
-        evaluation_id: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def replay_policy_evaluation(
-        self,
-        *,
-        evaluation_id: str,
-        body: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def record_policy_evaluation_event(
-        self,
-        *,
-        evaluation_id: str,
-        body: dict[str, Any],
-        idempotency_key: str | None,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_policy_evaluation_lineage(
-        self,
-        *,
-        evaluation_id: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_policy_sign_off_package(
-        self,
-        *,
-        evaluation_id: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_policy_evaluation_workflow(
-        self,
-        *,
-        evaluation_id: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def record_policy_sign_off_decision(
-        self,
-        *,
-        evaluation_id: str,
-        body: dict[str, Any],
-        idempotency_key: str | None,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def request_policy_report_package(
-        self,
-        *,
-        evaluation_id: str,
-        body: dict[str, Any],
-        idempotency_key: str | None,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def request_policy_ai_evidence(
-        self,
-        *,
-        evaluation_id: str,
-        body: dict[str, Any],
-        idempotency_key: str | None,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-
-class ProposalClient(Protocol):
-    async def simulate_proposal(
-        self,
-        body: dict[str, Any],
-        idempotency_key: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def create_proposal_artifact(
-        self,
-        body: dict[str, Any],
-        idempotency_key: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def create_proposal(
-        self,
-        body: dict[str, Any],
-        idempotency_key: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def list_proposals(
-        self,
-        params: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_proposal(
-        self,
-        proposal_id: str,
-        include_evidence: bool,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_proposal_version(
-        self,
-        proposal_id: str,
-        version_no: int,
-        include_evidence: bool,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def create_proposal_version(
-        self,
-        proposal_id: str,
-        body: dict[str, Any],
-        idempotency_key: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def create_proposal_async(
-        self,
-        body: dict[str, Any],
-        idempotency_key: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def create_proposal_version_async(
-        self,
-        proposal_id: str,
-        body: dict[str, Any],
-        idempotency_key: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_proposal_operation(
-        self,
-        operation_id: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_proposal_operation_by_correlation(
-        self,
-        operation_correlation_id: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_proposal_operation_replay_evidence(
-        self,
-        operation_id: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def transition_proposal(
-        self,
-        proposal_id: str,
-        body: dict[str, Any],
-        idempotency_key: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def record_approval(
-        self,
-        proposal_id: str,
-        body: dict[str, Any],
-        idempotency_key: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_workflow_events(
-        self,
-        proposal_id: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_approvals(
-        self,
-        proposal_id: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_proposal_lineage(
-        self,
-        proposal_id: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_proposal_version_replay_evidence(
-        self,
-        proposal_id: str,
-        version_no: int,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_proposal_idempotency_record(
-        self,
-        idempotency_key: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def regenerate_proposal_narrative(
-        self,
-        proposal_id: str,
-        version_no: int,
-        body: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_proposal_narrative(
-        self,
-        proposal_id: str,
-        version_no: int,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def review_proposal_narrative(
-        self,
-        proposal_id: str,
-        version_no: int,
-        body: dict[str, Any],
-        idempotency_key: str | None,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def create_report_request(
-        self,
-        proposal_id: str,
-        body: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def create_execution_handoff(
-        self,
-        proposal_id: str,
-        body: dict[str, Any],
-        idempotency_key: str | None,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_delivery_summary(
-        self,
-        proposal_id: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_delivery_events(
-        self,
-        proposal_id: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_execution_status(
-        self,
-        proposal_id: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def record_execution_update(
-        self,
-        proposal_id: str,
-        body: dict[str, Any],
-        idempotency_key: str | None,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def create_proposal_memo(
-        self,
-        proposal_id: str,
-        version_no: int,
-        body: dict[str, Any],
-        idempotency_key: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_proposal_memo(
-        self,
-        proposal_id: str,
-        version_no: int,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_proposal_memo_projection(
-        self,
-        proposal_id: str,
-        version_no: int,
-        audience: str | None,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def review_proposal_memo(
-        self,
-        proposal_id: str,
-        version_no: int,
-        body: dict[str, Any],
-        idempotency_key: str | None,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def record_proposal_memo_report_package_event(
-        self,
-        proposal_id: str,
-        version_no: int,
-        body: dict[str, Any],
-        idempotency_key: str | None,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def request_proposal_memo_report_package(
-        self,
-        proposal_id: str,
-        version_no: int,
-        body: dict[str, Any],
-        idempotency_key: str | None,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def request_proposal_memo_ai_commentary(
-        self,
-        proposal_id: str,
-        version_no: int,
-        body: dict[str, Any],
-        idempotency_key: str | None,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_proposal_memo_lineage(
-        self,
-        proposal_id: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_proposal_memo_replay_evidence(
-        self,
-        proposal_id: str,
-        version_no: int,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
 class ArchiveDocumentClient(Protocol):
     async def get_document_metadata(
         self,
@@ -1224,67 +657,5 @@ class PerformanceWorkspaceCoreClient(Protocol):
         benchmark_currency: str | None = None,
         benchmark_status: str | None = "active",
         benchmark_type: str | None = "composite",
-    ) -> tuple[int, dict[str, Any]]: ...
-
-
-class AdvisorBriefAiClient(Protocol):
-    async def execute_workflow_pack(
-        self,
-        *,
-        pack_id: str,
-        version: str,
-        environment: str,
-        caller_identity_class: str,
-        workflow_surface: str | None,
-        task_request: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_observability_runtime_status(
-        self,
-        *,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_workflow_pack_run_consumer_view(
-        self,
-        *,
-        run_id: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_workflow_pack_run_operator_profile(
-        self,
-        *,
-        run_id: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def list_workflow_pack_task_flows(
-        self,
-        *,
-        correlation_id: str,
-        workflow_pack_id: str | None = None,
-        caller: str | None = None,
-        workflow_surface: str | None = None,
-        limit: int = 25,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def apply_workflow_pack_run_review_action(
-        self,
-        *,
-        run_id: str,
-        correlation_id: str,
-        request_payload: dict[str, Any],
-    ) -> tuple[int, dict[str, Any]]: ...
-
-
-class AdvisorBriefAdviseClient(Protocol):
-    async def get_platform_capabilities(
-        self,
-        *,
-        consumer_system: str = "lotus-gateway",
-        tenant_id: str = "default",
-        correlation_id: str,
     ) -> tuple[int, dict[str, Any]]: ...
